+++ Runtime/Events/StringEvent.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace SODD.Events
+{
+    [CreateAssetMenu(menuName = Framework.Events.String, fileName = nameof(StringEvent), order = Framework.MenuOrders.String)]
+    public sealed class StringEvent : Event<string>
+    {
+    }
+}