+++ CHANGELOG.md
@@ -19,3 +19,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - IntEventListener, a component that listens for scriptable object events with int payloads.
 - FloatEventListener, a component that listens for scriptable object events with float payloads.
 - BoolEventListener, a component that listens for scriptable object events with bool payloads.
+- StringEvent class, a scriptable object event that takes a string as payload.